replace an old avl_tree by new one
