+++ crates/avl_tree/Cargo.toml
@@ -1,12 +1,8 @@
 [package]
 name = "avl_tree"
 version = "0.1.0"
-edition = "2018"
+edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-
-[dev-dependencies]
-rand = "0.8.4"
-superslice = "1.0.0"
